refactor acoustic models
